cleaned warnings and major restructure to crate
