+++ bin/src/bin.rs
