+++ crates/mempool/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "mempool"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
