+++ crates/neuroxide/src/layers/linear.rs
@@ -4,7 +4,7 @@ use std::sync::{Arc, RwLock};
 use crate::ops::op_generic::Operation;
 use crate::ops::add::AddOp;
 use crate::ops::matmul::MatMulOp;
-use crate::types::T::TensorElement;
+use crate::types::t::TensorElement;
 use crate::types::{device::Device, tensor::Tensor, tensordb::TensorDB};
 use crate::utils::node_uid::make_node_uid;
 
@@ -14,7 +14,7 @@ pub struct Linear<T> {
     db: Arc<RwLock<TensorDB<T>>>,
     input_features: usize,
     output_features: usize,
-    useBias: bool,
+    use_bias: bool,
     pub weights: Arc<RwLock<Tensor<T>>>,
     pub bias: Option<Arc<RwLock<Tensor<T>>>>,
 }
@@ -35,7 +35,7 @@ where
             db: db.clone(),
             input_features,
             output_features,
-            useBias: use_bias,
+            use_bias,
             weights: Arc::new(RwLock::new(weights)),
             bias: bias.map(|b| Arc::new(RwLock::new(b))),
         }
