+++ crates/neuroxide/src/ops/cos.rs
@@ -1,21 +1,22 @@
 use crate::ops::op_generic::{Ops, Operation};
 use crate::types::device::Device;
 use crate::types::tensor::Tensor;
-use crate::types::T::TensorElement;
+use crate::types::t::TensorElement;
 use crate::utils::node_uid::make_node_uid;
+use cfg_if::cfg_if;
 
 
 #[cfg(feature = "cuda")]
-extern "C" {
-pub fn ln_kernel(len: i32, a: *mut f32, c: *mut*mut f32) -> CudnnStatusT;
+unsafe extern "C" {
+pub fn cos_kernel(len: i32, a: *mut f32, c: *mut*mut f32) -> CudnnStatusT;
 }
 
 pub type CudnnStatusT = i32; // usually cuDNN uses enums as return statuses
 
 #[derive(Debug, Clone)]
-pub struct LnOp;
+pub struct CosOp;
 
-impl<T> Operation<T> for LnOp
+impl<T> Operation<T> for CosOp
 where
     T: TensorElement
 {
@@ -23,19 +24,27 @@ where
         assert!(inputs.len() == 1);
         let result: Vec<T>; // = vec![T::default(); len as usize];
 
-        let mut cuda_ptr: Option<*mut f32> = None;
+        cfg_if! {
+            if #[cfg(feature = "cuda")] {
+                let mut cuda_ptr: Option<*mut f32> = None;
+            } else {
+                let cuda_ptr: Option<*mut f32> = None;
+            }
+        }
+
 
         match inputs[0].device {
             Device::CPU => {
-                result = inputs[0].data.iter().map(|x| x.ln()).collect();
+                result = inputs[0].data.iter().map(|x| x.cos()).collect();
             }
             Device::CUDA => {
+
                 #[cfg(feature = "cuda")]
                 unsafe {
                     let len = inputs[0].data.len() as i32;
                     let mut data: f32 = 0.0;
                     let mut ptr_to_data: *mut f32 = &mut data;
-                    ln_kernel(len, inputs[0].cuda_ptr.unwrap(), &mut ptr_to_data);
+                    cos_kernel(len, inputs[0].cuda_ptr.unwrap(), &mut ptr_to_data);
                     cuda_ptr = Some(ptr_to_data);
                     let r = vec![0.0f32; len as usize];
                     result = r.iter().map(|&x| <T as num::NumCast>::from(x).unwrap()).collect();
@@ -60,7 +69,7 @@ where
             data: result,
             shape: inputs[0].shape.clone(),
             device: inputs[0].device,
-            op: Ops::LnEnum,
+            op: Ops::CosEnum,
             requires_grad: inputs[0].requires_grad,
             op_chain: result_graph,
             op_head: result_id,
@@ -80,14 +89,14 @@ where
         assert!(inputs.len() == 1);
         let mut grad_data = vec![T::default(); inputs[0].data.len()];
         for i in 0..inputs[0].data.len() {
-            grad_data[i] = T::one()/inputs[0].data[i];
+            grad_data[i] = T::from(-1).unwrap() * inputs[0].data[i].sin();
         }
         Tensor {
             id: inputs[0].id,
             data: grad_data,
             shape: inputs[0].shape.clone(),
-            device: inputs[0].device,
-            op: Ops::LnEnum,
+            device: Device::CPU,
+            op: Ops::CosEnum,
             requires_grad: inputs[0].requires_grad,
             op_chain: inputs[0].op_chain.clone(),
             op_head: inputs[0].op_head,
@@ -96,4 +105,3 @@ where
         }
     }
 }
-
