+++ crates/neuroxide/src/ops/pow.rs
@@ -3,14 +3,15 @@ use petgraph::prelude::GraphMap;
 use crate::ops::op_generic::{Ops, Operation};
 use crate::types::device::Device;
 use crate::types::tensor::Tensor;
-use crate::types::T::TensorElement;
+use crate::types::t::TensorElement;
 use crate::utils::array_utils::broadcast_shapes_linear;
 use crate::utils::node_uid::make_node_uid;
+use cfg_if::cfg_if;
 
 
 
 #[cfg(feature = "cuda")]
-extern "C" {
+unsafe extern "C" {
 pub fn pow_kernel(len: i32, a: *mut f32, b: *mut f32, c: *mut*mut f32) -> CudnnStatusT;
 }
 
@@ -40,7 +41,13 @@ where
         let final_shape: Vec<usize> = a_arr.shape().iter().map(|x| *x as usize).collect();
         
         let result: Vec<T>; // = vec![T::default(); len as usize];
-        let mut cuda_ptr: Option<*mut f32> = None;
+        cfg_if! {
+            if #[cfg(feature = "cuda")] {
+                let mut cuda_ptr: Option<*mut f32> = None;
+            } else {
+                let cuda_ptr: Option<*mut f32> = None;
+            }
+        }
 
         match inputs[0].device {
             Device::CPU => {
