+++ crates/neuroxide/src/utils/errors.rs
