+++ crates/neuroxide/src/utils/node_uid.rs
@@ -2,5 +2,5 @@ use rand::Rng;
 
 
 pub fn make_node_uid() -> i32 {
-    return rand::thread_rng().r#gen()
+    return rand::rng().random()
 }
